Extract ExitTool class
